backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

onnxruntime==1.18.1
numpy
Pillow==10.4.0

pydantic==2.7.4
pydantic-settings==2.3.4